refactoring: move log and config to common package level
