+++ cmd/import_businesses.go
@@ -10,9 +10,9 @@ import (
 	"hynie.de/ohmab/ent/business"
 	_ "hynie.de/ohmab/ent/runtime"
 	schemas "hynie.de/ohmab/ent/schema"
-	"hynie.de/ohmab/internal/pkg/config"
+	"hynie.de/ohmab/internal/pkg/common/config"
+	"hynie.de/ohmab/internal/pkg/common/log"
 	"hynie.de/ohmab/internal/pkg/db"
-	"hynie.de/ohmab/internal/pkg/log"
 	"hynie.de/ohmab/internal/pkg/privacy"
 	"hynie.de/ohmab/internal/pkg/utils"
 	"os"
