+++ cmd/import_timetables.go
@@ -12,9 +12,9 @@ import (
 	schemas "hynie.de/ohmab/ent/schema"
 	"hynie.de/ohmab/ent/schema/constants"
 	"hynie.de/ohmab/ent/timetable"
-	"hynie.de/ohmab/internal/pkg/config"
+	"hynie.de/ohmab/internal/pkg/common/config"
+	"hynie.de/ohmab/internal/pkg/common/log"
 	"hynie.de/ohmab/internal/pkg/db"
-	"hynie.de/ohmab/internal/pkg/log"
 	"hynie.de/ohmab/internal/pkg/privacy"
 	"os"
 	"strconv"
