+++ cmd/server.go
@@ -17,9 +17,9 @@ import (
 	_ "hynie.de/ohmab/ent/runtime"
 	"hynie.de/ohmab/ent/schema"
 	"hynie.de/ohmab/ent/user"
-	"hynie.de/ohmab/internal/pkg/config"
+	"hynie.de/ohmab/internal/pkg/common/config"
+	"hynie.de/ohmab/internal/pkg/common/log"
 	"hynie.de/ohmab/internal/pkg/db"
-	"hynie.de/ohmab/internal/pkg/log"
 	"hynie.de/ohmab/internal/pkg/privacy"
 	"hynie.de/ohmab/internal/pkg/routes"
 	"math/rand"
