+++ internal/pkg/routes/auth.go
@@ -8,8 +8,8 @@ import (
 	"github.com/google/uuid"
 	"hynie.de/ohmab/ent"
 	"hynie.de/ohmab/ent/user"
-	"hynie.de/ohmab/internal/pkg/config"
-	"hynie.de/ohmab/internal/pkg/log"
+	"hynie.de/ohmab/internal/pkg/common/config"
+	"hynie.de/ohmab/internal/pkg/common/log"
 	"hynie.de/ohmab/internal/pkg/privacy"
 	"hynie.de/ohmab/internal/pkg/utils"
 	"net/http"
